cancel-tests: only clean up when cancelling enqueued jobs (not running ones)

+++ server/autotest_enqueuer.py
@@ -135,10 +135,11 @@ def cancel_test(markus_address, run_ids, **kw):
                 job = rq.job.Job.fetch(job_id)
             except rq.exceptions.NoSuchJobError:
                 return
-            files_path = job.kwargs['files_path']
-            if files_path:
-                shutil.rmtree(files_path, onerror=ats.ignore_missing_dir_error)
-            job.cancel()
+            if job.is_queued():
+                files_path = job.kwargs['files_path']
+                if files_path:
+                    shutil.rmtree(files_path, onerror=ats.ignore_missing_dir_error)
+                job.cancel()
 
 def get_schema(**kw):
     """
